requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Hyperparameter optimization
optuna>=2.10.0  # Optional
hyperopt>=0.2.7  # Optional

# Deployment
flask>=2.0.0  # For API deployment
streamlit>=1.10.0  # For dashboard deployment

# Utilities
pyyaml>=5.4.0
joblib>=1.1.0
python-dotenv>=0.19.0 